+++ xcube_geodb/sql/geodb--0.1.8.sql
@@ -1,7 +1,7 @@
 CREATE SCHEMA IF NOT EXISTS geodb_user_info;
 
 
-CREATE SEQUENCE IF NOT EXISTS geodb_user_info.seq_geodb_user_info_id
+CREATE SEQUENCE IF NOT EXISTS public.geodb_user_info_id_seq
     INCREMENT 1
     START 2
     MINVALUE 1
@@ -9,13 +9,91 @@ CREATE SEQUENCE IF NOT EXISTS geodb_user_info.seq_geodb_user_info_id
     CACHE 1;
 
 
-CREATE TABLE IF NOT EXISTS "geodb_user_info"."geodb_user_info"(
-    id INT NOT NULL PRIMARY KEY DEFAULT nextval('geodb_user_info.seq_geodb_user_info_id'),
+CREATE TABLE IF NOT EXISTS public."geodb_user_info"(
+    id INT NOT NULL PRIMARY KEY DEFAULT nextval('geodb_user_info_id_seq'),
     user_name CHARACTER VARYING (255) NOT NULL UNIQUE,
     start_date DATE NOT NULL,
     subscription TEXT NOT NULL,
     permissions TEXT NOT NULL
 );
+
+
+CREATE OR REPLACE FUNCTION public.geodb_register_user_trg_func()
+    RETURNS trigger
+    LANGUAGE 'plpgsql'
+    COST 100
+    VOLATILE NOT LEAKPROOF
+AS $BODY$
+BEGIN
+        IF NEW.user_name IS NOT NULL THEN
+            EXECUTE format('SELECT geodb_register_user(''%s''::text, ''bla''::text)', NEW.user_name);
+        END IF;
+
+        RETURN NEW;
+    END;
+$BODY$;
+
+
+CREATE TRIGGER geodb_register_user_trg
+    AFTER INSERT ON "geodb_user_info"
+    FOR EACH ROW
+    EXECUTE PROCEDURE geodb_register_user_trg_func();
+
+
+CREATE SEQUENCE public.geodb_user_databases_seq
+    INCREMENT 1
+    START 1
+    MINVALUE 1
+    MAXVALUE 9223372036854775807
+    CACHE 1;
+
+
+CREATE TABLE public.geodb_user_databases
+(
+    id bigint NOT NULL DEFAULT nextval('geodb_user_databases_seq'::regclass),
+    name character varying COLLATE pg_catalog."default" NOT NULL,
+    owner character varying COLLATE pg_catalog."default" NOT NULL,
+    CONSTRAINT geodb_user_databases_pkey PRIMARY KEY (id),
+    CONSTRAINT unique_db_name_owner UNIQUE (name, owner)
+)
+WITH (
+    OIDS = FALSE
+)
+TABLESPACE pg_default;
+
+CREATE FUNCTION public.geodb_create_database(
+	database text)
+    RETURNS void
+    LANGUAGE 'plpgsql'
+
+    COST 100
+    VOLATILE
+AS $BODY$
+DECLARE usr text;
+BEGIN
+    usr := (SELECT geodb_whoami());
+
+    EXECUTE format('INSERT INTO geodb_user_databases(name, owner) VALUES(''%s'', ''%s'')
+				   ', "database", usr);
+END
+$BODY$;
+
+
+CREATE FUNCTION public.geodb_truncate_database(
+	database text)
+    RETURNS void
+    LANGUAGE 'plpgsql'
+
+    COST 100
+    VOLATILE
+AS $BODY$
+DECLARE usr text;
+BEGIN
+    usr := (SELECT geodb_whoami());
+
+    EXECUTE format('DELETE FROM geodb_user_databases WHERE name=''%s'' and owner=''%s''', "database", usr);
+END
+$BODY$;
 
 
 CREATE FUNCTION public.geodb_add_properties(IN collection text, IN properties json)
@@ -176,13 +254,19 @@ END
 $BODY$;
 
 
-CREATE FUNCTION public.geodb_list_collections()
+CREATE FUNCTION public.geodb_get_my_collections(
+    "database" TEXT DEFAULT NULL
+)
     RETURNS TABLE(src json)
     LANGUAGE 'plpgsql'
 AS $BODY$
 DECLARE usr text;
 BEGIN
     usr := (SELECT geodb_whoami());
+    IF database IS NULL
+    THEN
+        database := usr;
+    END IF;
 
     RETURN QUERY EXECUTE format('SELECT JSON_AGG(src) as js ' ||
                                 'FROM (SELECT
@@ -191,8 +275,35 @@ BEGIN
                                            pg_catalog.pg_tables
                                         WHERE
                                            schemaname = ''public''
-                                            AND tableowner = ''%s'') AS src',
-                                usr, usr);
+                                           AND tablename LIKE ''%s_%%''
+                                           AND tableowner = ''%s'') AS src',
+                                "database", "database", usr);
+
+END
+$BODY$;
+
+
+CREATE FUNCTION public.geodb_list_grants(
+	database TEXT
+	)
+    RETURNS TABLE(src json)
+    LANGUAGE 'plpgsql'
+
+    COST 100
+    VOLATILE
+    ROWS 1000
+AS $BODY$
+DECLARE usr text;
+BEGIN
+    usr := (SELECT geodb_whoami());
+
+    RETURN QUERY EXECUTE format('SELECT JSON_AGG(src) as js
+                                FROM (SELECT
+								regexp_replace(table_name, ''%s_'', '''') as table_name,
+								grantee
+                                FROM information_schema.role_table_grants
+                                WHERE grantor = ''%s'' AND grantee != ''%s'') AS src',
+                                database, usr, usr);
 
 END
 $BODY$;
@@ -201,18 +312,20 @@ $BODY$;
 CREATE FUNCTION public.geodb_list_grants()
     RETURNS TABLE(src json)
     LANGUAGE 'plpgsql'
+
+    COST 100
+    VOLATILE
+    ROWS 1000
 AS $BODY$
 DECLARE usr text;
 BEGIN
     usr := (SELECT geodb_whoami());
-    RETURN QUERY EXECUTE format('SELECT JSON_AGG(src) as js ' ||
-                                'FROM (SELECT table_name, grantee ' ||
-                                'FROM information_schema.role_table_grants ' ||
-                                'WHERE grantor = ''%s'' AND grantee != ''%s'') AS src',
-                                usr, usr);
+
+    RETURN QUERY SELECT geodb_list_grants(usr);
 
 END
 $BODY$;
+
 
 
 CREATE FUNCTION public.geodb_rename_collection(
@@ -431,43 +544,6 @@ BEGIN
 	RETURN QUERY EXECUTE format('SELECT geodb_get_user_usage(''%I'', TRUE)', me);
 END
 $BODY$;
-
-CREATE FUNCTION public.geodb_get_my_collections()
-    RETURNS TABLE(src json)
-    LANGUAGE 'plpgsql'
-
-    COST 100
-    VOLATILE
-    ROWS 1000
-AS $BODY$
-DECLARE me TEXT;
-BEGIN
-    SELECT geodb_whoami() INTO me;
-    RETURN QUERY SELECT geodb_get_my_collections(me);
-END
-$BODY$;
-
-
-CREATE FUNCTION public.geodb_get_my_collections(user_name text)
-    RETURNS TABLE(src json)
-    LANGUAGE 'plpgsql'
-
-    COST 100
-    VOLATILE
-    ROWS 1000
-AS $BODY$
-BEGIN
-    RETURN QUERY EXECUTE format('SELECT JSON_AGG(vals) as src FROM (
-		SELECT table_name, grantor as database, grantee FROM information_schema.table_privileges
-			WHERE table_schema = ''public''
-			AND (grantee = ''PUBLIC'' OR grantee = ''%I'')
-			AND grantor <> ''rdsadmin''
-			AND grantor <> ''postgres''
-            AND privilege_type = ''SELECT''
-		) as vals', user_name);
-END
-$BODY$;
-
 
 
 CREATE OR REPLACE FUNCTION public.geodb_get_pg(